requirements.txt
discord.py
Flask
python-dotenv
SQLAlchemy
psycopg2
stripe
pika
cryptography
pyngrok
gunicorn